Break out night watch examples into table and single select folders.
