+++ packages/terra-table/tests/nightwatch/single-select-table/components/SingleRowSelectableTable.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Table from '../../../lib/Table';
+import Table from '../../../../lib/Table';
 
 const SingleRowSelectableTable = () => (
   <Table isStriped={false}>
@@ -9,19 +9,9 @@ const SingleRowSelectableTable = () => (
       <Table.HeaderCell content={'Phone Number'} key={'PHONE_NUMBER'} minWidth={'large'} />
     </Table.Header>
     <Table.SingleSelectableRows>
-      <Table.Row isSelected key={'PERSON_0'}>
+      <Table.Row key={'PERSON_0'}>
         <Table.Cell content="John Smith" key="NAME" />
         <Table.Cell content="123 Adams Drive" key="ADDRESS" />
-        <Table.Cell content="111-222-3333" key="PHONE_NUMBER" />
-      </Table.Row>
-      <Table.Row key={'PERSON_1'}>
-        <Table.Cell content="Jane Smith" key="NAME" />
-        <Table.Cell content="321 Drive Street" key="ADDRESS" />
-        <Table.Cell content="111-222-3333" key="PHONE_NUMBER" />
-      </Table.Row>
-      <Table.Row key={'PERSON_2'}>
-        <Table.Cell content="Dave Smith" key="NAME" />
-        <Table.Cell content="213 Raymond Road" key="ADDRESS" />
         <Table.Cell content="111-222-3333" key="PHONE_NUMBER" />
       </Table.Row>
     </Table.SingleSelectableRows>
