+++ packages/terra-table/tests/nightwatch/single-select-table/components/SingleRowSelectableTableOnChange.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Table from '../../../lib/Table';
+import Table from '../../../../lib/Table';
 
 class SingleRowSelectableTable extends React.Component {
   constructor(props) {
