+++ packages/terra-table/tests/nightwatch/table/TableTests.jsx
@@ -13,12 +13,7 @@ const TableTests = () => (
       <li><Link to="/tests/table-tests/table-no-rows">Table - No Rows</Link></li>
       <li><Link to="/tests/table-tests/table-highlighted-rows">Table - Highlighted Rows</Link></li>
       <li><Link to="/tests/table-tests/table-sort-indicator">Table - Sort Indicator on Header</Link></li>
-      <li><Link to="/tests/table-tests/selectable-table">Selectable Table - Default</Link></li>
-      <li><Link to="/tests/table-tests/selectable-table-single-row">Selectable Table - Single Row</Link></li>
-      <li><Link to="/tests/table-tests/selectable-table-no-rows">Selectable Table - No Rows</Link></li>
-      <li><Link to="/tests/table-tests/selectable-table-onchange">Selectable Table - OnChange</Link></li>
       <li><Link to="/tests/table-tests/table-subheaders">Table - Subheaders</Link></li>
-      <li><Link to="/tests/table-tests/table-selectable-subheaders">Selectable Table - Subheaders</Link></li>
     </ul>
   </div>
 );
