+++ packages/terra-table/tests/nightwatch/table/components/TableNoRows.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Table from '../../../lib/Table';
+import Table from '../../../../lib/Table';
 
 const SingleRowTable = () => (
   <Table isStriped={false}>
